feat: use SocketIO instead of BufferedRWPair
Performance is suffering from buffering because of all the python copy. Since we know the size of the message we can avoid reading into a buffer and only reading the size we need

This will do 2 more syscalls but less internal buffering so its up to 50% faster since the syscalls should are far faster than the python code.