fix: pass return value to SendReply.__exit__
To return an error in a ServiceInterface, you raise a DBusError which
is caught and translated to a message and sent over the bus. However,
the context manager that catches this exception was incorrectly not
sending a return value of True when the exception was handled. This
caused the _message_reader() returned by build_message_reader() to
log the error even though it was handled.