feat: don't import backends by default
BREAKING CHANGE:
Since the top-level `__init__.py` is always imported when any subpackage
is imported, this caused both backends (glib and aio) to always be
imported. This is not desirable, since it means that the glib backend
will always be imported, even if the user only wants to use the aio.
This wastes resources and in some cases can even cause a crash.

Fixes: https://github.com/Bluetooth-Devices/dbus-fast/issues/242