chore: alternative name owner tracking defeat to appease Cython
Cython demands that a dict-typed object really be a dict and not a
subclass, presumably because it wants to shortcut the virtual dispatch
and implement all the methods in native code. Thus, my first attempt at
defeating the name owner tracking in MessageBus was a non-starter. This
commit switches to a different strategy: pre-filling the mapping so that
get_interface() won't try to shoot off a GetNameOwner call to the bus.