feat!: use tuple for STRUCT types instead of list
Return tuple instead of list for STRUCT types when unmarshaling D-Bus
messages. Marshalling still accepts both lists and tuples for STRUCT
types.

This is a breaking change for anyone who was relying on STRUCT types
being returned as lists. Operations like indexing and unpacking will
still work the same, so hopefully the impact is minimal.

The motivation behind this is to allow proper Python type hints in
`dbus_fast`. Python lists can only have a single type for all elements
which is not compatible with D-Bus STRUCT types. Python tuples, on the
other hand, exactly match the D-Bus STRUCT semantics.