feat: rename method and signal decorators to dbus_method and dbus_signal
Rename the decorators for defining D-Bus methods and signals from
`@method` and `@signal` to `@dbus_method` and `@dbus_signal`
respectively.

This makes it consistent with the `@dbus_property` decorator used for
defining D-Bus properties and makes it clearer that these decorators are
specific to D-Bus functionality.

The old decorator names are preserved so that this does not break
existing codebases.

Closes: https://github.com/Bluetooth-Devices/dbus-fast/issues/318