fix: raise ValueError if the same interface instance is exported multiple times
Add a check to ServiceInterface._add_bus() to raise a ValueError if the
same interface instance has already been exported on the bus. The
interfaces are tracked in a set, so exporting them multiple times will
cause the errors later when trying to unexport them.

This can have a side effect of causing bus.wait_for_disconnect() to hang
forever, because the later exception prevented the bus finalizer from
completing the future that wait_for_disconnect() was waiting on.

It will be better to raise the error early to make it easier to find
and fix the problem.