chore: always wait for disconnect in async tests
Add a wait for disconnect for a buses used in async tests to ensure
that all resources are cleaned up before the test ends. This helps to
catch more bugs in tests that would otherwise be overlooked. And also
avoids lots of warnings about unclosed resources when running the
tests.