fix: move call to close socket
Move the call for closing the D-Bus socket from `dbus_fast.aio.MessageBus.disconnect()`
to `dbus_fast.BaseMessageBus._finalize()`. The socket it opened by
`BaseMessageBus`, so it should be closed by it as well. Also,
`disconnect()` may not always be explicitly called, e.g. when connecting
to the bus, so we need this to be in the `_finalize()` to ensure that
it is always closed.