chore: recommend returning multiple values as tuples instead of lists
In Python, the natural way to return multiple values from a function is
as a tuple. Change the docs to recommend returning tuples instead of lists
for D-Bus methods and signals that return multiple values.

Tests are updated as well to show that tuples are preferred (and that
this is not a breaking change).